requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
openai==1.12.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
requests==2.31.0
pandas==2.2.1
numpy==1.26.4
pytest==8.0.2
black==24.2.0
flake8==7.0.0 